+++ decisionserver/decisionserverruntime/script/runds.sh
@@ -1,5 +1,4 @@
 #!/bin/bash
-set -x
 if [ ! -f /config/initialized.flag ] ; then
 	cd /config/apps/DecisionService.war/WEB-INF;
 	sed -i $'/<\/web-app>/{e cat /config/basicAuth.xml\n}' web.xml
@@ -10,6 +9,15 @@ if [ ! -f /config/initialized.flag ] ; then
 	sed -i '\#<config-property-name>DefaultConnectionManagerProperties#,\#<config-property-value/># s|<config-property-value/>|<config-property-value>pool.maxSize='$CONNECTION_POOL_SIZE',pool.waitTimeout=3000</config-property-value>|' ra.xml;
 	touch /config/initialized.flag
 fi;
+
+# Begin - Configuration for the tls security
+if [ -n "$KEYSTORE_PASSWORD" ]
+then
+	sed -i 's|__PASSWORD__|'$KEYSTORE_PASSWORD'|g' /config/tlsSecurity.xml
+else
+	sed -i 's|__PASSWORD__|'password'|g' /config/tlsSecurity.xml
+fi
+# End - Configuration for the tls security
 
 if [ -n "$DECISIONSERVERCONSOLE_NAME" ]
 then
@@ -23,13 +31,17 @@ then
 fi
 
 # Begin - Configuration for the user registry
-if [ "$REGISTRY" = "ldap" ]
+# For kubernetes in the case of a user want to override the configuration
+if [ ! -f "/config/webSecurity.xml" ]
 then
-	echo "Use LDAP registry"
-	cp /config/webSecurity-ldap.xml /config/webSecurity.xml
-else
-	echo "Use basic registry"
-	cp /config/webSecurity-basic.xml /config/webSecurity.xml
+	if [ "$REGISTRY" = "ldap" ]
+	then
+		echo "Use LDAP registry"
+		cp /config/webSecurity-ldap.xml /config/webSecurity.xml
+	else
+		echo "Use basic registry"
+		cp /config/webSecurity-basic.xml /config/webSecurity.xml
+	fi
 fi
 # End - Configuration for the user registry
 
@@ -40,37 +52,37 @@ then
 	wget -nv $DB_DRIVER_URL
     case $DB_DRIVER_URL in
     	*derby* ) rm /config/resources/derby*
-				  mv derby* /config/resources 
-				  cp /config/datasource-derby.xml /config/datasource.xml 
+				  mv derby* /config/resources
+				  cp /config/datasource-derby.xml /config/datasource.xml
 				  ;;
       	*mysql* ) rm /config/resources/mysql*
-				  mv mysql* /config/resources 
-				  cp /config/datasource-mysql.xml /config/datasource.xml 
+				  mv mysql* /config/resources
+				  cp /config/datasource-mysql.xml /config/datasource.xml
 				  ;;
       	*postgres* ) rm /config/resources/postgres*
-					 mv postgres* /config/resources 
-					 cp /config/datasource-postgres.xml /config/datasource.xml 
-					 ;; 
+					 mv postgres* /config/resources
+					 cp /config/datasource-postgres.xml /config/datasource.xml
+					 ;;
 		*db2* ) rm /config/resources/db2*
-				mv db2* /config/resources 
-				cp /config/datasource-db2.xml /config/datasource.xml 
-				;; 
+				mv db2* /config/resources
+				cp /config/datasource-db2.xml /config/datasource.xml
+				;;
 	esac
 elif [ -n "$DB_TYPE" ]
 then
 	echo "Use DB_TYPE: $DB_TYPE"
 	case $DB_TYPE in
 		*derby* ) if [ ! -f /config/resources/derby* ]; then  /script/installDerby.sh; fi
-				  cp /config/datasource-derby.xml /config/datasource.xml 
+				  cp /config/datasource-derby.xml /config/datasource.xml
 				  ;;
 		*mysql* ) if [ ! -f /config/resources/mysql* ]; then /script/installMySQL.sh; fi
-				  cp /config/datasource-mysql.xml /config/datasource.xml 
+				  cp /config/datasource-mysql.xml /config/datasource.xml
 				  ;;
 		# For postgreSQL, we do not have to install the driver here since it is installed by default at build time
-      	*postgres* ) cp /config/datasource-postgres.xml /config/datasource.xml 
+      	*postgres* ) cp /config/datasource-postgres.xml /config/datasource.xml
 					 ;;
 		# For DB2, we do not have to install the driver here since it is supposed to be provided through the drivers folder at build time
-		*db2* ) cp /config/datasource-db2.xml /config/datasource.xml 
+		*db2* ) cp /config/datasource-db2.xml /config/datasource.xml
 				;;
 	esac
 else
@@ -80,24 +92,24 @@ fi
 # End - Configuration for the database
 
 # Begin - Change values for the datasource if required
-if [ -n "$DB_SERVER_NAME" ] 
-then 
+if [ -n "$DB_SERVER_NAME" ]
+then
 	sed -i 's|dbserver|'$DB_SERVER_NAME'|g' /config/datasource.xml
 fi
-if [ -n "$DB_PORT_NUMBER" ] 
-then 
+if [ -n "$DB_PORT_NUMBER" ]
+then
 	sed -i 's|5432|'$DB_PORT_NUMBER'|g' /config/datasource.xml
 fi
-if [ -n "$DB_NAME" ] 
-then 
+if [ -n "$DB_NAME" ]
+then
 	sed -i 's|odmdb|'$DB_NAME'|g' /config/datasource.xml
 fi
-if [ -n "$DB_USER" ] 
-then 
+if [ -n "$DB_USER" ]
+then
 	sed -i 's|odmusr|'$DB_USER'|g' /config/datasource.xml
 fi
-if [ -n "$DB_PASSWORD" ] 
-then 
+if [ -n "$DB_PASSWORD" ]
+then
 	sed -i 's|odmpwd|'$DB_PASSWORD'|g' /config/datasource.xml
 fi
 # End - Change values for the datasource if required
